requirements.txt
mongoengine==0.9.0
pymongo==2.8
django==1.7
djangorestframework==3.1.3
django-rest-framework-mongoengine==2.0.2